+++ click_man/core.py
@@ -10,11 +10,11 @@ generate man pages for entire click applications.
 """
 
 import os
+from distutils.version import LooseVersion
 
 import click
 
 from .man import ManPage
-
 
 def generate_man_page(ctx, version=None):
     """
@@ -29,13 +29,15 @@ def generate_man_page(ctx, version=None):
     # Create man page with the details from the given context
     man_page = ManPage(ctx.command_path)
     man_page.version = version
-    man_page.short_help = ctx.command.short_help
+    man_page.short_help = ctx.command.get_short_help_str()
     man_page.description = ctx.command.help
     man_page.synopsis = ' '.join(ctx.command.collect_usage_pieces(ctx))
-    man_page.options = [x.get_help_record(None) for x in ctx.command.params if isinstance(x, click.Option)]
+    man_page.options = [x.get_help_record(ctx) for x in ctx.command.params if isinstance(x, click.Option)]
     commands = getattr(ctx.command, 'commands', None)
     if commands:
-        man_page.commands = [(k, v.short_help) for k, v in commands.items()]
+        man_page.commands = [
+            (k, v.get_short_help_str()) for k, v in commands.items()
+        ]
 
     return str(man_page)
 
@@ -63,4 +65,9 @@ def write_man_pages(name, cli, parent_ctx=None, version=None, target_dir=None):
 
     commands = getattr(cli, 'commands', {})
     for name, command in commands.items():
+        if LooseVersion(click.__version__) >= LooseVersion("7.0"):
+            # Since Click 7.0, we have been able to mark commands as hidden
+            if command.hidden:
+                # Do not write a man page for a hidden command
+                continue
         write_man_pages(name, command, parent_ctx=ctx, version=version, target_dir=target_dir)