+++ CHANGELOG.md
@@ -6,6 +6,22 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 *Nothing here yet*
 
+## [v0.4.0]
+- Discover command if not in entry_points
+
+## [v0.3.0]
+- Make compatible with click >= 7.0
+
+## [v0.2.2]
+- Release
+
+## [v0.2.1]
+### Added
+- Add information to man page for commands. Refs #3
+
+### Fixed
+- Use hyphens to split words in a command. Refs #2
+
 ## [v0.2.0]
 ### Added
 - Allow to specify version via command line
@@ -26,7 +42,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Please also see `git log`
 
-[Unreleased]: https://github.com/timofurrer/click-man/compare/v0.2.0...HEAD
-[v0.2.0]: https://github.com/timofurrer/click-man/compare/v0.1.1...v0.2.0
-[v0.1.1]: https://github.com/timofurrer/click-man/compare/v0.1.0...v0.1.1
-[v0.1.0]: https://github.com/timofurrer/click-man/compare/30626839cc048856f799eb0bcd9e731fff4221dc...v0.1.0
+[Unreleased]: https://github.com/click-contrib/click-man/compare/v0.4.0...HEAD
+[v0.4.0]: https://github.com/click-contrib/click-man/compare/v0.3.0...v0.4.0
+[v0.3.0]: https://github.com/click-contrib/click-man/compare/v0.2.2...v0.3.0
+[v0.2.2]: https://github.com/click-contrib/click-man/compare/v0.2.1...v0.2.2
+[v0.2.1]: https://github.com/click-contrib/click-man/compare/v0.2.0...v0.2.1
+[v0.2.0]: https://github.com/click-contrib/click-man/compare/v0.1.1...v0.2.0
+[v0.1.1]: https://github.com/click-contrib/click-man/compare/v0.1.0...v0.1.1
+[v0.1.0]: https://github.com/click-contrib/click-man/compare/30626839cc048856f799eb0bcd9e731fff4221dc...v0.1.0